Corrected an issue where the global offset would not be included into the header, causing errors

+++ code/board/Core/Src/lib/maths/maths.c
@@ -51,7 +51,7 @@ inline int Double2ADC(float val, uint16_t *code)
  */
 inline int S162Double(int8_t code, float *val)
 {
-	*val = code / 64.0;
+	*val = (float)code / 64.0;
 	return 0;
 }
